requirements.txt
fastapi[standard]
uvicorn
pydantic
python-dotenv
requests
groq